requirements.txt
psycopg2
SQLAlchemy
Flask-SQLAlchemy
flask-restful